+++ packages/iov-bcp/src/connection.ts
@@ -3,7 +3,7 @@ import { ReadonlyDate } from "readonly-date";
 import { As } from "type-tagger";
 import { Stream } from "xstream";
 
-import { PostableBytes } from "./codec";
+import { PostableBytes, TxReadCodec } from "./codec";
 import {
   Address,
   Amount,
@@ -188,6 +188,20 @@ export interface BlockchainConnection {
   // Synchonous side-effect free getters are implemented as properties
   /** The ID of the chain we are connected to */
   readonly chainId: ChainId;
+  /**
+   * The codec used by this connection to interact with the blockchain.
+   *
+   * This allows you to perform address operations on a connection
+   * instance:
+   *
+   * ```
+   * // validate
+   * const valid = connection.codec.isValidAddress(userInput);
+   * // convert
+   * const senderAddress = connection.codec.identityToAddress(sender);
+   * ```
+   */
+  readonly codec: TxReadCodec;
 
   // blockchain
   readonly disconnect: () => void;